+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/actions/FlexibleSearchExecuteQueryAction.kt
@@ -1,7 +1,7 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2014-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,19 +18,23 @@
  */
 package com.intellij.idea.plugin.hybris.flexibleSearch.actions
 
-import com.intellij.icons.AllIcons
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.flexibleSearch.file.FlexibleSearchFileType
 import com.intellij.idea.plugin.hybris.tools.remote.action.AbstractExecuteAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
 
-class FlexibleSearchExecuteQueryAction : AbstractExecuteAction(
-    message("hybris.fxs.actions.execute_query"),
-    message("hybris.fxs.actions.execute_query.description"),
-    AllIcons.Actions.Execute
-) {
+class FlexibleSearchExecuteQueryAction : AbstractExecuteAction() {
+
+    init {
+        with (templatePresentation) {
+            text = message("hybris.fxs.actions.execute_query")
+            description = message("hybris.fxs.actions.execute_query.description")
+            icon = HybrisIcons.CONSOLE_EXECUTE
+        }
+    }
 
     override val extension = FlexibleSearchFileType.instance.defaultExtension
     override val consoleName = HybrisConstants.FLEXIBLE_SEARCH_CONSOLE_TITLE