+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExExecuteAction.kt
@@ -17,18 +17,22 @@
  */
 package com.intellij.idea.plugin.hybris.impex.actions
 
-import com.intellij.icons.AllIcons
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.impex.file.ImpexFileType
 import com.intellij.idea.plugin.hybris.tools.remote.action.AbstractExecuteAction
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 
-class ImpExExecuteAction : AbstractExecuteAction(
-    message("hybris.impex.actions.execute_query"),
-    message("hybris.impex.actions.execute_query.description"),
-    AllIcons.Actions.Execute
-) {
+class ImpExExecuteAction : AbstractExecuteAction() {
+
+    init {
+        with (templatePresentation) {
+            text = message("hybris.impex.actions.execute_query")
+            description = message("hybris.impex.actions.execute_query.description")
+            icon = HybrisIcons.CONSOLE_EXECUTE
+        }
+    }
 
     override fun getActionUpdateThread() = ActionUpdateThread.EDT
     override val extension = ImpexFileType.INSTANCE.defaultExtension