+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added new toolbar action `Remove Column` [#554](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/554)
 - Ignore case for type in the reference type (e.g. `groups(customer.uid)`) [#545](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/545)
 - Added `Execute` and `Validate` actions to the toolbar [#553](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/553)
+- Added `Select Statement` action to the toolbar [#556](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/556)
 
 ### `CockpitNG` enhancements
 - Added code completion for AdvancedSearch `operator` parameter [#537](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/537)